Fix some memory leaks in the widget handling code.
They only manifest themselves in case ParsePreferenceItem() fails: in
this case, DictionaryValue::Set() is not called and |preference| leaks.

Fix it by turning most raw pointers in WidgetHandler::Parse() into
scoped_ptr's, even where it is not strictly necessary.

WidgetInfo::Set()'s prototype had to be adjusted a little to accept and
forward scoped_ptr's instead of base::Value*'s, but it is a change for
the good that prevents similar mistakes in the future.

CID=323513
CID=323514