Finish fixing test failures caused by Chromium bug 501171.
The fix for Linux browser tests was not enough, not even for Linux: the
observers must be registered with content::RenderThread before
EnsureWebKitInitialized() is called, otherwise they will not be notified
of the Blink's initialization.

The most direct consequence is that the SetNetworkAvailableTest Android
test fails because webkit_initialized_ is always false in
XWalkRenderProcessObserver.

Instead of adapting to https://codereview.chromium.org/1373923005, base
the code on https://codereview.chromium.org/1382983002 which is the
proper fix up there too.

BUG=XWALK-6047