Actually adapt to https://codereview.chromium.org/1373763003
Adapting to that change is not as simple as changing the return value of
the call to base::JSONReader::Read() in XWalkContent::SetManifest().

In order to produce a base::DictionaryValue for the Manifest class, we
were going through a few layers of indirection and scoped_ptrs. By
making |manifest_value| a scoped_ptr, we added another layer which
resulted in some double deletions in the end (|manifest_value| would
delete |manifest_dictionary|, and so would Manifest's destructor).

This fixes Android tests such as LoadAppFromManifestTest.