[Android] Implement XWalkContentLifecycleNotifier.
This patch provides a way to observe XWalkView lifecycle on the java
side. By implementing XWalkContentLifecycleNotifier.Observer and
registering with the notifier, clients will get notified when there
are alive instances of XWalkViews (i.e. xwalk_content) or when all
XWalkView instances have been destroyed.

This is useful for APIs which can be disabled when there are no
living XWalkViews, e.g. Network Information API.

This code was ported from upstream,
CL: https://codereview.chromium.org/1419083010

This is Part 1, the observer should also be implemented for XWalkView
lifecycle observers.

BUG=XWALK-5892, XWALK-6178