xesh: Get it to a compilable state.
xesh is not been built as part of the "xwalk" or "xwalk_builder"
targets, which means it has been bitrotting for a few years.

This patch adapts the code to several changes that have happened to the
Chromium code base in the meantime (`CommandLine` needs the `base`
namespace, removal of `base::MessageLoopProxy` etc).

It is not enough to get it to run without crashing though, so someone
else needs to fix it or remove it from the tree altogether.

BUG=XWALK-6219