[Android] Rewrite generate_version_code.py.
Instead of having part of the logic in xwalk.gyp and part in the Python
script, do everything in the latter, while also taking advantage of the
fact that we've already know Crosswalk's version number by the time we
invoke the script, so we do not need to read it again.

Also update the comment describing how we derive the version code
number, as it was a bit outdated (we do not reset the build number when
we bump either the major or the minor number).