[Lite] Move compressed library file into lib folder
In Lite, the compressed library is placed in /res/raw as a resource
file. it is named as libxwalkcore.so.x86 or libxwalkcore.so.armeabi_v7a.
This cause several problem:
1) In a bundled aar or webview.zip, these two files are placed together,
   Android regards them as duplicated resources.
2) The gradle of latest version regards it as illegal resource name.
3) Gradle cannot build multiple apks because it cannot filter resource
   files for different abi target.

In this PR, we don't put it res/raw any more, we rename it, put it in
lib/x86/libxwalkcoreCompressed.so (x86 for example) and modify accordingly
in make_apk.py and in java class which handles decompressing in runtime.

One WEAKNESS: During installation, Android will copy the compressed lib
 into a read only folder, which occupy more flash of the phone.

BUG=XWALK-6379