extensions: Port away from deprecated V8 APIs.
In preparation for Chromium M49 (which will enable stricter compilation
flags for more code by default), port the extensions subsystem away from
deprecated V8 API calls. They will be removed soon, and it is better to
avoid being surprised late in the game.

While here, enable the `chromium_code' variable in `extensions.gyp` so
that we build all targets there (currently, only `xwalk_extensions`)
with more compiler warnings and errors so that we catch uses of
deprecated APIs earlier in the future.

BUG=XWALK-6276
BUG=XWALK-6372