[Android] Change color of root layer to avoid white screen on startup
If set the WindowBackground of Activity to RED, and webpage page's backgroudn to RED,
when application starts or resumes after click back key, it will showup a white screen
shortly. We can change the first screen to a fix color by setting the root layer color
of ContentViewCore. This can help to avoid white screen.

BUG=XWALK-4809,XWALK-4995