build: Disable use_sysroot by default.
This is a no-op in M48, but starting with M49 `use_sysroot` is enabled
by default on Linux. This means the build will use a Debian Wheezy
chroot to build and use as few system dependencies as possible.

It is a nice idea, but we cannot use it with Crosswalk at the moment
because we have an additional dependency on libnotify that is not part
of the Wheezy chroot shipped by Chromium.

BUG=XWALK-5806