[Android] Fix the failed android tests on M49.
For OnReceivedErrorTest#testCacheMiss, the related resource id was
removed in
https://github.com/crosswalk-project/crosswalk/commit/29838edb08b71e49c43eff580ac02f4709d1cfaa, this leads to the description is null.

For SetInitialScaleTest#testSetInitialScale1, the clobber initial scale
should works in quirks mode, please refer to https://codereview.chromium.org/67193002,
but crosswalk has no API to set quirks mode. Meanwhile the previous
CriteriaHelper.poll* did not check the result inside function, it was fixed in
https://codereview.chromium.org/1491063002, then initial scale issue was
exposed.

For OnPageFinishedTest#testOnPageFinishedNotCalledForJavaScriptUrl, it
was caused by https://codereview.chromium.org/1304373007, the return
value of GetLoadingUrl() was reset to data url instead of unreachable
url.

BUG=XWALK-6533