[Android] Enable pull-to-refresh
With this enabled, when user pull from the page top, page will be reloaded.
This can be disabled by flag --disable-pull-to-refresh-effect.
Pull-to-refresh differs from edge effect. For the edge effect, it is from
android and it "performs the graphical effect used at the edges of scrollable
widgets when the user scrolls beyond the content bounds in 2D space.".
Detailed documents is here:
http://developer.android.com/reference/android/widget/EdgeEffect.html.
As to the pull-to-refresh, it is from google chrome. It means reload the page
when user vertically overscrolling content at page top.

BUG=XWALK-6277