sysapps: Use real Promises instead of a polyfill.
Back when the sysapps code was being written, V8's and Chromium's
Promises implementation was still under development, and only reached
Crosswalk months later.

The sysapps code resorted to using a very basic polyfill at the
time (which has a different syntax and is not as asynchronous as a
proper implementation), and nobody updated the code once we were able to
use Chromium's Promises implementation.

This change finally does it: instead requiring the "sysapps_promise"
module, we now just use Chromium's Promise object, reducing the
dependency on the third-party `common_promise_api.js`.