[Android] Update warning in setAcceptFileSchemeCookies's API documentation
Since M46, there is a change in upstream to match the comments
in CookieManager "Note that calls to this method will have no
effect if made after a WebView or CookieManager instance has
been created". For XWalkCookieManager, the function can't be
defined static for reflection limitation, so the comment is
not correct.

BUG=XWALK-6274