[Android] Implement the feature about custom url scheme.
This patch is to support the custom url scheme, e.g. "whatsapp://",
"intent://".
For "intent://", if the related app, e.g. Barcode Scanner, was
installed, then start the activity to launch this scanner, if it was
not installed, XWalkView will load the fallback url if it was specified.
For "whatsapp://", if whatsapp was installed, start the activity to
launch it, if it was not installed, then try to install it from market
according to the package name. If the package name is null, will report
"ERR_UNKNOWN_URL_SCHEME".

BUG=XWALK-6690