[XWALK-6262] Refactor to remove duplicated code after Presentation API is re-implemented on Android
This is a leftover task of XWALK-4646.

Prior to XWALK-4646, there was the impl. of W3C presentation API for Windows platform. It was Windows only impl at that time.

After some experiment, there was a common subset identified in source code between Android & Windows platform impl., but considering the fact that XWALK-4646 was in high priority (P1 blocker) at that time, and I was unable to obtain a Windows box to test the modified behavior on Windows, I made a decision to copy the common code and modify part of it to Android-specific source code.

After XWALK-4646 is resolved, this XWALK-6262 was opened to track the leftover task, which is to share the common code between Windows & Android platform.

This task was delayed due to Chromium 49 rebase (change of content interface that we're relying on and behavior slightly), and also delayed a couple of weeks due to my IDF'16 trip.

Now this task is finally done.