[Windows] Fix notifications on Windows.
Make sure to set the AppUserModelID correctly so that it matches
what app-tools is setting on the shortcut. This will fix the bug
where packaged app would not show notifications : this was due to
some very old code inside xwalk setting AppUserModelID to the
application_id. This code is not needed anymore because we're going
to set AppUserModelID to the application name being launched. This
patch will also fix the grouping in the notification center on builds
leading to Anniversary Update because instead of showing the app_id
it will show the actual name of the app.

BUG=XWALK-4990