[Android] Enable the wide viewport quirk by default.
This patch is to enable wide viewport quirk by default.
The WideViewport quirk is to maintain compatibility with Android apps,
and is required so long as XWalkSettings.{get|set}UseWideViewPort is
supported.
Without enabling, the page is zoom out which is different from WebView.
Porting the test case for wide viewport from Android WebView.
Fixed the failed case of setInitialScale().

Please refer to https://chromiumcodereview.appspot.com/23496047
for more details.

BUG=XWALK-6762