Fix race condition when RenderProcess or ExtensionProcess dies.
When either of these two processes are dying then it may be that there
are still messages being filtered in the IPC channel. If so then the filter
may have dangling pointers to the extension thread server or the ui thread
server (as those have been deleted already). The solution is to invalidate
the filter when deleting the XWalkExtensionData object. Please note that
this commit partially revert XWALK-4295 beside the call to RemoveFilter in
OnRenderProcessHostClosed which was the cause of the crash (double delete
because RemoveFilter also delete the filter).

BUG=XWALK-6856