Refactor to remove duplicated code after Presentation API is re-implemented on Android
Prior to XWALK-4646, there was the impl. of W3C presentation API for Windows platform. It was Windows-only impl at that time.

After some experiment, there was a common subset identified in source code between Android & Windows platform impl., but considering the fact that XWALK-4646 was in high priority (P1 blocker) at that time, and I was unable to obtain a Windows box to test the modified behavior on Windows, I made a decision to copy the common code and modify part of it to Android-specific source code.

After XWALK-4646 is resolved, this XWALK-6262 was opened to track the leftover task, which is to share the common code between Windows & Android platform. It was delayed due to rebase and other resource issues. Now it's finally done.

BUG=XWALK-6262