[Windows] Download support
Implement download support for Windows as it was already almost finished.
I removed the code when not built with Aura because it's dead code
(Windows, Linux are built with Aura by default). I also put the code
to figure out the Download path in the common place with other key paths.

BUG=XWALK-4969