[Android] Indicate that a malformed URL isn't accepted by XWalkCooikeManager
Some customers complain long debug time to know that url without scheme is invalid
when using XWalkCooikeManager, which isn't required in Android WebView.

This is because CookieManager already use WebAddress to add scheme into the url.
For the different design, xwalk will not do the same implementation, but it's better
to throw an exception to user in this case.

BUG=XWALK-6766
(cherry picked from commit 426e3cb0dee1fb3a95f709ec6d28779072049a44)