[Android] Fix crash when setting surface type through layout
XWalkViewBridge will try to get the member mAnimatable from XWalkView.
But if the class type in the layout file is XWalkView's subclass, the
member can't be found and a exception is thrown. This patch will look
for the member by going up the class hierarchy.
(cherry picked from commit cbcfbb271088ca0276f3221ca10d8b6098203e69)