[Linux] Fix the crash in ColorChooser
The original implementation with CreateWindowWithContext may cause
X receive a bad window parameter and the ColorChooserView is placed
at a fixed position but not bundled with the root window.
Now we use CreateWindowWithParent to parent the ColorChooserView with
toplevel window to avoid the X error mentioned above and the view is
bundled with the root window and it can be moved together with the root
window.

BUG=XWALK-6958