[Android] Refactor XWalkView to receive events directly.
In Crosswalk, ContentView and ContentViewRenderView were used to render
page, some events were dispatched from ContentView to XWalkView, such as
touching and clicking. Its inefficiency and in latest version of
Chromium, ContentView was declared as private, so it should be
deprecated.

Now XWalkViewInternal implements all of the related interfaces and
overrides all of the functions which are used in ContentView, so
XWalkViewInternal could be regarded as another ContentView.
ContentViewRenderView and XWalkViewInternal were added to XWalkView
and XWalkViewInternal could receive the events directly.
Also fix some bugs.

BUG=XWALK-6118