Fix the video play issue after M51 rebase.
This issue was caused by https://codereview.chromium.org/1577433004
The condition of 'OS=="android"' was removed, this leads to
'proprietary_codecs' was always 0, then mp4 and other codecs could not
be added to related 'MediaFormatMappings'.

'proprietary_codecs' was defined in the condition in src/build/common.gypi,
so in src/xwalk/common.gypi, 'proprietary_codecs' should also be moved
to condition to keep same scope level.

BUG=XWALK-7032