[Android] Fix the potential vulnerability issue.
For SSL errors, if user allows it, native layer will keep the host and
the error num, the future communications with same host and error will
not trigger the onReceivedSslError callback. Even if the certificate is
invalid, all of requests will also be passed, its vulnerable for some
serious SSL error.
Especially with MITM, the sensitive messages will be stolen, e.g.
passwords, credit cards.
Deny the invalid requests for some serious SSL error to avoid this issue.

For keeping the host and error num, please refer to
https://codereview.chromium.org/794023002
We also have the same implementation.

BUG=XWALK-6986