[Windows][Linux] Fix Geolocation crashing after the user accept the prompt.
Bug is a regression after M51 and incorrect patch of
8d0c7a657433574edc397e2dbebe8dcc690e45c1 which removes the functions
CreateRequestContext and CreateRequestContextForStoragePartition from
XWalkContentBrowserClient as they moved to BrowserContext (so
XWalkBrowserContext). The problem is that they were setting the
url_request_context_getter_ member before calling XWalkBrowserContext.
The URL request context is then used afterwards inside xwalk to query
things like geolocation. With that accidental removal it was always null
so it was crashing/asserting later in the geolocation code. Fix is to
setup a getter from the browser context to access the URL request context
when creating the AccessTokenStore.

BUG=XWALK-7067