[Windows] Presentation API : Fix bug where no screens would be listed as available.
When creating the monitors list we need to make sure that they are
initialized first as not in use. In order to do that we're removing the
user defined DisplayInfo constructor and thanks to C++11 the field in_use
will be set to false. Before the field was not initialized by default
and was being affected a random value (which then resolve as true) and
all monitors becomes in use and the API never let you choose one.

BUG=XWALK-7098