Enable audio playback with enable-unified-media-pipeline flag.
The 'ffmpeg_branding' was defined in 'third_party/ffmpeg/ffmpeg.gyp',
defaut value is 'chromium'. It was overridden in 'xwalk/build/common.gypi',
but the overridden value did not work since this variable coming from
different file was not at the same scope level.
Move 'ffmpeg_branding' to 'variables.conditions' to keep the same scope
level instead of 'variables.variables'.

BUG=XWALK-6874