Add a copy constructor to XWalkRuntimeFeatures::Feature.
This fixes a build error on Linux when `use_sysroot` is enabled;
something in the old libstdc++ shipped in the Debian Wheezy sysroot
triggers Chromium's chromium-style clang plugin.

While the error message says this is actually caused by a constructor
having an inlined body, this isn't really the case: both constructors
previously declared were not inlined in the class declaration. The copy
constructor was implicitly added as an inline member, though.

```
../../xwalk/runtime/common/xwalk_runtime_features.h:34:10: error: [chromium-style] Complex constructor has an inlined body.
  struct Feature {
         ^
```

BUG=XWALK-7022