Android: Add --use-srcjars to reflection_generator.py.
This is the biggest commit in the recent refactorings of this script.
--use-srcjars will cause the files generated by reflection_generator to
be written into .srcjar files (namely, `bridge.srcjar` and
`wrapper.srcjar`) instead of into the file system. A .srcjar file is
just a zip file containing .java files.

At the moment this has no effect, as the option is disabled by default
and not enabled in gyp, but it will be used in GN, whose
`android_library()` template accepts an `srcjars` argument with exactly
the kind of files we are producing. Using srcjars allows us to properly
specify our dependencies in GN, as the files generated by
reflection_generator.py cannot be specified directly as dependencies to
another target.

For more information, see the `java_cpp_enum.py` script in Chromium and
https://groups.google.com/a/chromium.org/forum/#!topic/gn-dev/GokFp6TorgI

BUG=XWALK-3674