[Android] Rewrite generate_xwalk_core_library.py
Finally do something I've wanted for years: a completely rewrite of the
horrendous `generate_xwalk_core_library.py`, with an eye on making it
easier for us to move to GN.

Even though this commit changes a lot of lines, it is basically just
simplifying the script by making it dumber. All the file names and paths
have been moved to where they should have been in the first place: the
build system. We now properly depend on all targets that produce files
and directories we use, and pass them to the script via arguments. The
script is now only responsible for creating the right directories and
copying things around (doing processing on resources in some cases).

One of the largest changes is in `CopyResources()`, which still does a
non-trivial amount of work. It still works on pre-processed resource zip
files produced by other targets, but we embed the information about
where the original sources are (for 9-patch images) in the build system
and pass it to the script. Once we fully switch to GN, we'll be able to
infer the source directories from the zip file path.