[Android] Rewrite generate_app_packaging_tool.py
Also delegate most decisions about file paths to the build system, and
only use the script to actually performs the copies and renamings
necessary to get the template directory looking like we want it to.

This simplifies the script and also paves the way for it to be used by
GN, which has different paths for certain files consumed by the script.