gn: Implement the xwalk_app_template target
This is fairly simple after commit cadbd11 ("[Android] Rewrite
generate_app_packaging_tool.py"): we just need to depend on the
recently-added `xwalk_core_library` and `xwalk_shared_library` targets
and specify the paths to everything we want to copy.

We also leverage GN's build_configs to determine the path to the runtime
JAR without having to hardcode it anywhere.