gn: Add the xwalk_core_library_docs target.
Unfortunately we still need to manually list all the files we want to
process as per
https://lists.crosswalk-project.org/pipermail/crosswalk-dev/2016-September/003927.html

We cannot call `javadoc` directly in GN, so a small wrapper,
`generate_javadoc.py`, has been added. It also knows how to deal with
`.srcjar` files, which is how the reflection wrapper code is currently
packaged.

The gyp version of the target has also been changed to use the wrapper
so that the targets in both build systems do not differ too much.

There is no `xwalk_core_library_java_app_part` target in GN, so we use
`xwalk_core_java` instead (and adjust gyp accordingly): the `app_part`
target was just merging `xwalk_core_java.jar` and the JARs it depends
on, but the documentation only needs the classes from
`xwalk_core_java.jar` itself, so there is no difference in the output.