gn: Use the AndroidManifest.xml from the template when creating AARs
We were previously creating the Android AARs using the
`AndroidManifest.xml` specified via `invoker.android_manifest`.

While that is correct for all other usages in the
`generate_android_project` template, using it in the AARs we create is
wrong, as the manifest specified via `invoker.android_manifest` is
generally the one in `//xwalk/runtime/android/core_internal_empty`
instead of the one
`//xwalk/build/android/xwalk{core,_shared}_library_template`. The ones
in `build/` specify a different package name and don't have an
`<application>` tag. This tag conflicts with the ones in the manifest
generated by Cordova when building an app, meaning Android GN builds
were not working with Cordova at all.

BUG=XWALK-7348