gn: Make xwalk_shared_library depend on additional resource targets
It is not clear if this is a bug or due to some odd design, but the
xwalk_shared_library Android project needs resources from both
`runtime/android/core_internal` (due to its usage of the reflection
layer wrapper classes) as well as resources from Chromium itself that
the core_internal code depends on.

These same dependencies are also part of `XWalkRuntimeLib` itself, so it
means we need to duplicate them.

For now, introduce a new variable, `core_internal_java_chromium_deps` to
make it clear that this is hopefully something provisional and add some
informative comments instructing people to not add non-//xwalk
dependencies to `xwalk_core_internal_java`s `deps` list.

BUG=XWALK-7350
BUG=XWALK-7358
BUG=XWALK-7366