gn: Copy //xwalk/VERSION to the build dir for generate_crosswalk_win_zip
When GN support for Crosswalk on Windows was landed, I thought there
would be no need to mimic what gyp does and copy //xwalk/VERSION to
`$root_out_dir`.

It turns out this is necessary because `generate_crosswalk_zip.py`
assumes all paths it stores in the zip archive are inside the build
directory. This means the zip file being generated with GN so far had a
relative path inside for `VERSION` (`../../xwalk/VERSION`) which did not
play well with `unzip` and other tools that need to extract it.

A sanity check has been added to `generate_crosswalk_zip.py` to make
sure this does not happen in the future.

While here, drop the `is_win` check from the
`generate_crosswalk_win_zip` target: it just needlessly complicates the
code with a check and another level of indentation.

BUG=XWALK-7336