Depend on blink, not blink_headers, in the extensions target
blink_headers is supposed to be used when targets only need access to
Blink structs and enums, which is not the case here, as
`xwalk_v8tools_module.cc` actually calls Blink code.

This was preventing `xwalk_extensions_unittest` from linking in debug
mode with GN and `is_component = true`. gyp was not affected because the
`xwalk_extensions` target there is always built as a static library, but
the code was updated there too for completeness.